pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-lexenv"
version = "0.1.0"
description = "LexEnv - AI Legal Document Analysis Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
inference = [
    "openai>=1.0.0",
    "google-generativeai>=0.8.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["lexenv", "lexenv.data", "server"]
package-dir = { "lexenv" = "lexenv", "lexenv.data" = "lexenv/data", "server" = "server" }
